docs and nits
