+++ crates/essential-rest-client/src/node_client.rs
@@ -1,13 +1,12 @@
-// #![deny(missing_docs)]
-
+use crate::handle_response;
 use essential_types::{
     contract::Contract, convert::bytes_from_word, predicate::Predicate, Block, ContentAddress, Key,
     Value, Word,
 };
-use reqwest::{Client, ClientBuilder, Response};
+use reqwest::{Client, ClientBuilder};
 use std::ops::Range;
 
-/// Client library for sending requests to the Essential node.
+/// Client that binds to an Essential node address.
 pub struct EssentialNodeClient {
     /// Async reqwest client to make requests with.
     client: Client,
@@ -26,24 +25,27 @@ impl EssentialNodeClient {
     /// Get contract at content address.
     pub async fn get_contract(&self, contract_ca: &ContentAddress) -> anyhow::Result<Contract> {
         let url = self.url.join(&format!("/get-contract/{contract_ca}"))?;
-        let response = handle_error(self.client.get(url).send().await?).await?;
+        let response = handle_response(self.client.get(url).send().await?).await?;
         Ok(response.json::<Contract>().await?)
     }
 
     /// Get predicate at content address.
     pub async fn get_predicate(&self, predicate_ca: &ContentAddress) -> anyhow::Result<Predicate> {
         let url = self.url.join(&format!("/get-predicate/{predicate_ca}"))?;
-        let response = handle_error(self.client.get(url).send().await?).await?;
+        let response = handle_response(self.client.get(url).send().await?).await?;
         Ok(response.json::<Predicate>().await?)
     }
 
     /// List blocks in the given L2 block number range.
+    ///
+    /// Blocks are only created if there are valid solutions.
+    /// Blocks are created on a regular interval.
     pub async fn list_blocks(&self, range: Range<Word>) -> anyhow::Result<Vec<Block>> {
         let url = self.url.join(&format!(
             "/list-blocks?start={}&end={}",
             range.start, range.end
         ))?;
-        let response = handle_error(self.client.get(url).send().await?).await?;
+        let response = handle_response(self.client.get(url).send().await?).await?;
         Ok(response.json::<Vec<Block>>().await?)
     }
 
@@ -56,11 +58,17 @@ impl EssentialNodeClient {
             "/list-contracts?start={}&end={}",
             range.start, range.end
         ))?;
-        let response = handle_error(self.client.get(url).send().await?).await?;
+        let response = handle_response(self.client.get(url).send().await?).await?;
         Ok(response.json::<Vec<(Word, Vec<Contract>)>>().await?)
     }
 
     /// Query state in the given contract address and key.
+    ///
+    /// This is the main way the front end application will interact with state.
+    /// It only really makes sense to query state where you know what the ABI of the contract is.
+    /// The state that's returned is a list of words of variable size.
+    /// The keys are also variable sized lists of words.
+    /// To make use of this API you need to know what type of contract you are querying.
     pub async fn query_state(
         &self,
         contract_ca: ContentAddress,
@@ -71,17 +79,7 @@ impl EssentialNodeClient {
         let url = self
             .url
             .join(&format!("/query-state/{contract_ca}/{key}"))?;
-        let response = handle_error(self.client.get(url).send().await?).await?;
+        let response = handle_response(self.client.get(url).send().await?).await?;
         Ok(response.json::<Option<Value>>().await?)
-    }
-}
-
-async fn handle_error(response: Response) -> anyhow::Result<Response> {
-    let status = response.status();
-    if status.is_success() {
-        Ok(response)
-    } else {
-        let text = response.text().await?;
-        Err(anyhow::anyhow!("{}: {}", status, text))
     }
 }
